# Optimized PlainHasher hashing. Trie insertions are >15% faster