--- a/util/plain_hasher/Cargo.toml
+++ b/util/plain_hasher/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "plain_hasher"
+description = "Hasher for 32-bit keys."
+version = "0.1.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+license = "MIT"
+keywords = ["hash", "hasher"]
+categories = ["no-std"]
+homepage = "https://github.com/paritytech/plain_hasher"
+
+[dependencies]
+crunchy = "0.1.6"