`PlainHasher::write` before:

```
test write ... bench:      18,025 ns/iter (+/- 1,425)
```

`PlainHasher::write` after:

```
test write ... bench:          58 ns/iter (+/- 10)
```


before:

```
test trie_insertions_32_mir_1k      ... bench:   3,956,331 ns/iter (+/- 577,529)
test trie_insertions_32_ran_1k      ... bench:   4,009,316 ns/iter (+/- 567,403)
test trie_insertions_random_mid     ... bench:   2,900,765 ns/iter (+/- 283,349)
test trie_insertions_six_high       ... bench:   2,981,152 ns/iter (+/- 331,667)
test trie_insertions_six_low        ... bench:   5,828,811 ns/iter (+/- 593,180)
test trie_insertions_six_mid        ... bench:   3,889,366 ns/iter (+/- 110,769)
test trie_iter                      ... bench:   6,755,724 ns/iter (+/- 1,199,892)
```

after:

```
test trie_insertions_32_mir_1k      ... bench:   3,262,684 ns/iter (+/- 118,147)
test trie_insertions_32_ran_1k      ... bench:   3,290,951 ns/iter (+/- 78,756)
test trie_insertions_random_mid     ... bench:   2,654,520 ns/iter (+/- 744,629)
test trie_insertions_six_high       ... bench:   2,492,217 ns/iter (+/- 236,359)
test trie_insertions_six_low        ... bench:   4,989,133 ns/iter (+/- 86,834)
test trie_insertions_six_mid        ... bench:   3,314,200 ns/iter (+/- 84,884)
test trie_iter                      ... bench:   6,749,411 ns/iter (+/- 6,490,631)
```