@debris 2017-08-17
I moved `PlainHasher` to a separate crate. here is the [source](https://github.com/debris/plain_hasher/blob/master/src/lib.rs)