@rphmeier 2017-08-18
indentation wrong here: https://github.com/debris/plain_hasher/blob/master/src/lib.rs#L45
BTW, shouldn't it be a `paritytech` crate -- @arkpar and I are also authors of PlainHasher