@debris 2017-08-21
@rphmeier updated readme and moved the repo to https://github.com/paritytech/plain_hasher

I will also add core-devs as owners on crates.io once they fix this bug: https://github.com/rust-lang/crates.io/issues/997