@arkpar 2017-08-28
I don't really see why `PlainHasher` needs to be on crates.io. It is too parity-specific to be generally useful.